Rename modal compoennt for search result
